+++ dqops/src/main/java/com/dqops/metadata/storage/localfiles/credentials/aws/AwsDefaultConfigProfileProvider.java
@@ -5,6 +5,7 @@ import com.dqops.core.secrets.SecretValueLookupContext;
 import com.dqops.metadata.storage.localfiles.credentials.DefaultCloudCredentialFileContent;
 import com.dqops.metadata.storage.localfiles.credentials.DefaultCloudCredentialFileNames;
 import com.dqops.utils.exceptions.DqoRuntimeException;
+import org.apache.parquet.Strings;
 import org.springframework.stereotype.Component;
 import software.amazon.awssdk.profiles.Profile;
 import software.amazon.awssdk.profiles.ProfileFile;
@@ -25,12 +26,13 @@ public class AwsDefaultConfigProfileProvider {
     /**
      * Provides the AWS config profile file.
      * @param secretValueLookupContext Secret value lookup context used to access shared credentials.
+     * @param profile AWS profile name.
      * @return AWS default config profile created from the credential files available in DQOps shared credentials.
      */
-    public static Optional<Profile> provideProfile(SecretValueLookupContext secretValueLookupContext){
+    public static Optional<Profile> provideProfile(SecretValueLookupContext secretValueLookupContext, String profile){
 
         String sectionName = "profiles";
-        String sectionTitle = "default";
+        String sectionTitle = Strings.isNullOrEmpty(profile) ? "default" : null;
 
         FileContent fileContent = CredentialsFileProvider.getCredentialFileContent(
                 DefaultCloudCredentialFileNames.AWS_DEFAULT_CONFIG_NAME,
@@ -41,13 +43,15 @@ public class AwsDefaultConfigProfileProvider {
             String keyContent = fileContent.getTextContent();
 
             if (Objects.equals(keyContent.replace("\r\n", "\n"), DefaultCloudCredentialFileContent.AWS_DEFAULT_CONFIG_INITIAL_CONTENT)) {
-                ProfileFile profileFile;
-                try (InputStream keyReaderStream = new ByteArrayInputStream(keyContent.getBytes(StandardCharsets.UTF_8))) {
-                    profileFile = ProfileFile.builder().content(keyReaderStream).type(ProfileFile.Type.CONFIGURATION).build();
-                    return profileFile.getSection(sectionName, sectionTitle);
-                } catch (IOException e) {
-                    throw new DqoRuntimeException("The .credentials/" + DefaultCloudCredentialFileNames.AWS_DEFAULT_CONFIG_NAME +
-                            " file contains default (fake) credentials. Please update the file by setting valid AWS default credentials.");
+                if (!Objects.equals(sectionTitle, "default")) {
+                    ProfileFile profileFile;
+                    try (InputStream keyReaderStream = new ByteArrayInputStream(keyContent.getBytes(StandardCharsets.UTF_8))) {
+                        profileFile = ProfileFile.builder().content(keyReaderStream).type(ProfileFile.Type.CONFIGURATION).build();
+                        return profileFile.getSection(sectionName, sectionTitle);
+                    } catch (IOException e) {
+                        throw new DqoRuntimeException("The .credentials/" + DefaultCloudCredentialFileNames.AWS_DEFAULT_CONFIG_NAME +
+                                " file contains default (fake) credentials. Please update the file by setting valid AWS default credentials.");
+                    }
                 }
             } else {
                 ProfileFile profileFile;