+++ CHANGELOG.md
@@ -1,3 +1,3 @@
 # 1.11.5
 
-
+* Support the default authentication chain to access AWS S3